Fix sizing for types.h - long long *should* be 64 bits

+++ src/types.h
@@ -1,17 +1,17 @@
 #ifndef AAA_TYPES_H
 #define AAA_TYPES_H
 
-typedef unsigned char   u8;
-typedef signed char     s8;
+typedef unsigned char       u8;
+typedef signed char         s8;
 
-typedef unsigned short  u16;
-typedef signed short    s16;
+typedef unsigned short      u16;
+typedef signed short        s16;
 
-typedef unsigned int    u32;
-typedef signed int      s32;
+typedef unsigned int        u32;
+typedef signed int          s32;
 
-typedef unsigned long   u64;
-typedef signed int      s64;
+typedef unsigned long long  u64;
+typedef signed long long    s64;
 
 // results are strictly a signed 32-bit integer
 typedef s32 result_t;
